Lift out string parsing to a dedicated function
This separates the parsing logic from the file-specific error
handling that only needs to be present when an actual on-disk file is
being parsed.

This also renames the grammar rule from "file" to "script" to better
represent what it actually is parsing (a series of
expressions in a given string).