Fully split out CLI and lib into separate crates within the workspace
The CLI has a number of dependencies that we don't want to export with
the library, notably notify and rand which require signficant extra
effort to configure in a webassembly context. Fully splitting this out
will separate their closures properly so we can depend on the library
crate in the sandbox without the extra baggage.